pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.3.7",
    "langchain>=1.2.0",
    "langchain-azure-ai>=1.0.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.4",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "pandas>=2.3.3",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
